requirements.txt
django==3.2.6
requests==2.26.0
djangorestframework==3.12.4
markdown==3.3.4
django-filter==21.1
django-ninja==0.16.2
